requirements.txt
fastapi==0.104.1
google-generativeai==0.8.3
httpx==0.25.2
python-docx==1.1.0
reportlab==4.0.7
python-dateutil==2.8.2
feedparser==6.0.10
requests==2.31.0
pydantic==2.5.1
uvicorn==0.24.0
